0.9.0

+++ WEB/PHP/拆分测试/数据库操作/新闻-添加&&删除/index_ok.php
@@ -4,24 +4,33 @@
     //连接服务器  
     include("conn.php");   //$conn = 服务器    
 
-    
-    $pid = $_POST['htmlID'];   //接收POST来的ID值        
-    
+    if($_POST['htmlID'] && $_POST['Submit']){
+            $pid = $_POST['htmlID'];   //接收POST来的ID值        
+        
+            // DELETE FROM `news` WHERE `news`.`id` = 6   数据库语法  
+            $del = mysqli_query($conn,"DELETE FROM news WHERE news.id='$pid' ");  
 
-    // DELETE FROM `news` WHERE `news`.`id` = 6   数据库语法  
-    $del = mysqli_query($conn,"DELETE FROM news WHERE news.id='$pid' ");  
-
-    if($del){
+            if($del){
+                echo "<script>
+                        alert('删除成功!');  
+                        window.location.href='index.php';  
+                    </script>";  
+            } else {
+                echo "<script>
+                        alert('删除失败!');  
+                        window.location.href='index.php';  
+                    </script>"; 
+            };
+    }; 
+    /*
+    else if($_POST['htmlID'] && $_POST['Submit2']){
         echo "<script>
-                alert('删除成功!');  
-                window.location.href='index.php';  
-            </script>";  
-    } else {
-        echo "<script>
-                alert('删除失败!');  
-                window.location.href='index.php';  
+                window.location.href='edit.php';  
             </script>"; 
-    }
+    };
+    */
+    
+    
 
 
 ?>
